Refactor part title and author method
