+++ src/blocks/utils/parseTitleAndAuthor.spec.ts
@@ -1,66 +1,4 @@
-type BookInfo = {
-  title: string;
-  author: string | undefined;
-};
-
-type Segment = {
-  text: string;
-  level: number;
-  startIndex: number;
-};
-
-export const parseTitleAndAuthor = (input: string): BookInfo => {
-  const stack: Segment[] = [];
-  let word = '';
-  let level = 0;
-  let lastStartIndex = 0;
-
-  for (let i = 0; i < input.length; i++) {
-    const char = input[i];
-
-    if (char === '(') {
-      if (level === 0) {
-        stack.push({ text: word.trim(), level, startIndex: lastStartIndex });
-        lastStartIndex = i;
-        word = '';
-      }
-      word += char;
-      level++;
-    } else if (char === ')') {
-      word += char;
-      if (level === 1) {
-        stack.push({ text: word.trim(), level, startIndex: lastStartIndex });
-        word = '';
-      }
-      level--;
-    } else {
-      word += char;
-    }
-  }
-
-  if (word.length > 0) {
-    stack.push({ text: word.trim(), level, startIndex: lastStartIndex });
-  }
-
-  console.log(stack.filter((s) => s.text.length > 0));
-
-  let author: string | undefined = undefined;
-
-  const hasAuthor = stack[stack.length - 1].level === 1;
-
-  if (hasAuthor) {
-    const lastSegment = stack.pop();
-    author = lastSegment?.text.replace(/^\(/, '').replace(/\)$/, '');
-  }
-
-  return {
-    title: stack
-      .map((s) => s.text)
-      .join(' ')
-      .trim(),
-    author,
-  };
-};
+import { parseTitleAndAuthor } from './parseTitleAndAuthor';
 
 describe('parseDateOfCreation', () => {
   test.each([
